backend/requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.1.0
python-decouple==3.8
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
reportlab==4.0.7
xlsxwriter==3.1.9
openpyxl==3.1.2
requests==2.31.0
cryptography==41.0.7
qrcode==7.4.2
python-barcode==0.15.1
bleach==6.1.0
django-storages==1.14.2
boto3==1.34.0
gunicorn==21.2.0
whitenoise==6.6.0
